+++ components/destination/DestinationHeader.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const DestinationHeader = () => {
+  return (
+    <div className="text-center mb-8">
+      <h1 className="text-4xl font-bold text-gray-900 mb-4">여행지 둘러보기</h1>
+      <p className="text-xl text-gray-600 max-w-2xl mx-auto">
+        전 세계의 아름다운 여행지를 발견하고 다음 여행을 계획해보세요
+      </p>
+    </div>
+  );
+};
+
+export default DestinationHeader;
