+++ utils/destination/constants.ts
@@ -1,0 +1,20 @@
+export const REGIONS = [
+  '전체',
+  '동아시아',
+  '동남아시아',
+  '유럽',
+  '북미',
+  '남미',
+  '오세아니아',
+  '아프리카',
+  '중동',
+];
+
+export const BUDGETS = ['전체', '저예산', '중간예산', '고예산'];
+
+export const BUDGET_THRESHOLDS = {
+  LOW: 200000,
+  HIGH: 900000,
+};
+
+export const POPULAR_THRESHOLD = 8;
